AinfluencerHub/AinfluencerHub/python/requirements.txt
# Server
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sse-starlette>=2.1.0,<3.0.0
python-multipart>=0.0.12,<1.0.0
pydantic>=2.7.0,<3.0.0

# Image / ML
Pillow>=10.0.0,<12.0.0
numpy>=1.24.0,<3.0.0
pyiqa>=0.1.12,<1.0.0

# HuggingFace / diffusion / training
diffusers>=0.30.0,<1.0.0
transformers>=4.51.0,<5.0.0
huggingface_hub>=0.23.0,<1.0.0
safetensors>=0.4.2,<1.0.0
accelerate>=0.30.0,<2.0.0
peft>=0.10.0,<1.0.0
bitsandbytes>=0.43.0,<1.0.0
torch>=2.1.0,<3.0.0
torchvision>=0.16.0,<1.0.0

# Face-consistent generation (IP-Adapter)
insightface>=0.7.3,<1.0.0
onnxruntime-gpu>=1.17.0,<2.0.0

# Video generation
imageio[ffmpeg]>=2.31.0,<3.0.0

# Utilities
requests>=2.31.0,<3.0.0
pyyaml>=6.0.1,<7.0.0

# Dev / test (safe to install everywhere; ruff/mypy/pytest are cheap)
pytest>=8.0.0,<9.0.0
httpx>=0.27.0,<1.0.0
ruff>=0.6.0,<1.0.0
mypy>=1.11.0,<2.0.0
